Rename lateral to PokedexSection
